renamed *Container, Cadastro components to Container*, CadastroProduto
